Have default stringifier for timestamps
